GamePlayPresenter: -Changed the if statement inside the submit method to remove the tiles placed on board when exchanging tiles from the bag -Replaced int tile size with double from higher precision in zoom out calculations. (tilesize returns a double) TileNode: -Second constructor of TileNode now accepts double in its parameters.

+++ Qwirkle_Beta_Version/src/qwirkle/view/gamePlayFrame/TileNode.java
@@ -22,7 +22,7 @@ public final class TileNode extends Rectangle {
         name = String.valueOf(hashCode());
     }
     //Computer Tile
-    public TileNode(Tile tile, int size) {
+    public TileNode(Tile tile, double size) {
         this.tile = tile;
         setWidth(size);
         setHeight(size);
